RDKEMW-9150:[SECVULN] - HTTPS support for firmware download- statered Fix

+++ src/rdkv_upgrade.c
@@ -513,6 +513,9 @@ int rdkv_upgrade_request(const RdkUpgradeContext_t* context, void** curl, int* p
 	    else if (ret_curl_code == RDKV_UPGRADE_ERROR_FORCE_EXIT) {
 		    return RDKV_UPGRADE_ERROR_FORCE_EXIT;
 	    }
+	    else if (ret_curl_code == RDKV_UPGRADE_ERROR_STATE_RED) {
+		    return RDKV_UPGRADE_ERROR_STATE_RED;
+	    }
 
             if (ret_curl_code != CURL_SUCCESS ||
                 (*pHttp_code != HTTP_SUCCESS && *pHttp_code != HTTP_CHUNK_SUCCESS && *pHttp_code != HTTP_PAGE_NOT_FOUND)) {
@@ -548,6 +551,9 @@ int rdkv_upgrade_request(const RdkUpgradeContext_t* context, void** curl, int* p
             }
             else if (ret_curl_code == RDKV_UPGRADE_ERROR_FORCE_EXIT) {
                     return RDKV_UPGRADE_ERROR_FORCE_EXIT;
+            }
+            else if (ret_curl_code == RDKV_UPGRADE_ERROR_STATE_RED) {
+                    return RDKV_UPGRADE_ERROR_STATE_RED;
             }
             if (ret_curl_code != CURL_SUCCESS ||
                 (*pHttp_code != HTTP_SUCCESS && *pHttp_code != HTTP_CHUNK_SUCCESS && *pHttp_code != HTTP_PAGE_NOT_FOUND)) {
@@ -1052,7 +1058,7 @@ int downloadFile(
             if (checkAndEnterStateRed(CURL_MTLS_LOCAL_CERTPROBLEM, disableStatsUpdate) != 0) {
                 SWLOG_ERROR("%s : State red entered due to MTLS cert problem\n", __FUNCTION__);
             }
-            return curl_ret_code;
+            return RDKV_UPGRADE_ERROR_STATE_RED;
         } else if (ret == STATE_RED_CERT_FETCH_FAILURE) {
             SWLOG_ERROR("%s : State red cert failed.\n", __FUNCTION__);
             return curl_ret_code;
@@ -1174,6 +1180,10 @@ int downloadFile(
         SWLOG_ERROR("%s : Direct Image upgrade Fail: curl ret:%d http_code:%d\n", __FUNCTION__, curl_ret_code, *httpCode);
         (server_type == HTTP_SSR_DIRECT) ? setDwnlState(RDKV_FWDNLD_DOWNLOAD_FAILED) : setDwnlState(RDKV_XCONF_FWDNLD_DOWNLOAD_FAILED);
         dwnlError(curl_ret_code, *httpCode, server_type,device_info,lastrun,disableStatsUpdate);
+        if (filePresentCheck(STATEREDFLAG) == 0) {
+            SWLOG_INFO("%s : State red entered during download, skipping retry\n", __FUNCTION__);
+            return RDKV_UPGRADE_ERROR_STATE_RED;
+        }
         if( *(file_dwnl.pathname) != 0 )
         {
             unlink(file_dwnl.pathname);
